Broke FrhodoDriver off into its own module
